Home

TestThings.jl

This package is purely for testing purposes, i.e. figuring out how Continuous Integration (with Travis and Appveyor), documentation (with Documenter.jl) and Coverage work.

The index contains all the functions in autodocs (maybe)

TestThings.bar

TestThings.foo

TestThings.foobar

TestThings.foobar

Autodocs contain all the doc-strings, nice!

TestThings.bar — Method.

bar(xs::Vector{<:Integer})

Like foo but instead returns the minimum of the values in xs, i.e.

\[bar(xs) = \{x \mid x \leq y \quad \forall y \in xs\}\]

source

TestThings.foo — Method.

foo(xs::Vector{<:Integer})

return the maximum of the values in xs, i.e.

\[foo(xs) = \{x \mid x \geq y \quad \forall y \in xs\}\]

source

TestThings.foobar — Method.

foobar(xs::Vector{<:Real})

returns the extrema of a vector of integers

source

TestThings.foobar — Method.

foobar(xs::Vector{<:Integer})

returns the extrema of a vector of integers

source